rename the module, move into a package
