+++ ldap-connection-manager/pom.xml
@@ -8,7 +8,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>ldap-connection-management</artifactId>
+    <artifactId>ldap-connection-manager</artifactId>
     <packaging>takari-jar</packaging>
 
     <properties>
