+++ ldap-connection-manager/src/main/java/com/walmartlabs/ollie/ldap/DomainControllerUrlResolver.java
@@ -1,4 +1,4 @@
-package com.walmartlabs.ollie;
+package com.walmartlabs.ollie.ldap;
 
 /*-
  * *****
