+++ ldap-connection-manager/src/main/java/com/walmartlabs/ollie/ldap/LdapConnectionManager.java
@@ -1,4 +1,4 @@
-package com.walmartlabs.ollie;
+package com.walmartlabs.ollie.ldap;
 
 /*-
  * *****
@@ -23,7 +23,6 @@ package com.walmartlabs.ollie;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.naming.AuthenticationException;
 import javax.naming.CommunicationException;
 import javax.naming.NamingException;
 import javax.naming.ldap.LdapContext;
